Synchronized movement
Two seperate entities are able to see one another move in real time properly now.

Look in Asset folder files for code written by Greg and Alex in regards to this commit.
Alex set up the PlayerSetup which allows the player to properly disable or enable the correct components so the game knows which parts it needs the network to transmit. Greg was able to set up the two needed components including the NetworkTransform and NetworkTransformChild.