Crosshair addition, camera fix
Added a crosshair for when the player is in the game. Was originally posting as soon as game started but now only spawns on the player instance in game. Fixed Camera view to assure that the graphics did not obscure the local player's view (such as the front metal pieces).
Removed a few standard assets that were unused to tidy up space.