Hosting/joining Games
Big additions made in JoinGame asset, HostGame asset, and ListRoomItem assets. these were made to handle the different functions of hosting and joining games on the server with others, where the ListRoomItem script is made to handle and give our server list that is shown on the GUI layout. Other small bug fixes were smashed during this process.